+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 - not yet released
 
 - add `:remove` method
+- add `:set_ttl` method
 
 ## 1.0.0 - 2022-08-21